refactor: combine preprocessor with conformer
Preprocessor and conformer were requiring many of the same attributes.
Instead of passing many of the attributes to one another combined the
classes. Preprocessor methods are private because they are more complex.

Also move some preprocessor functions to transformations
